requirements.txt
alembic==1.13.1
annotated-types==0.7.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.1.1
flask-sqlalchemy==3.1.1
flask==3.0.3
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.4
mako==1.3.5
markupsafe==2.1.5
psycopg2==2.9.9
pydantic-core==2.18.4
pydantic[email]==2.7.3
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.30
typing-extensions==4.12.2
urllib3==2.2.1
werkzeug==3.0.3
